update amount type

+++ BankingAPI.BLL/BankingWebAPI.BLL/Interface/IUserAccountRepository.cs
@@ -11,6 +11,6 @@ namespace BankingWebAPI.BLL.Interface;
     {
         Task<List<UserAccountDetail>> GetUSerAccountDetailsByUserIDRepositoryAsync(int userID);
         Task <APIResponseHandler<UserAccountDetail>> GetUserAccountDetailsByAccountNoRepositoryAsync(string accountNo);
-        Task <APIResponseHandler<long>> GetUserAccountBalanceByAccountNoAndUserIdRepositoryAsync(int userID,string accountNo);
+        Task <APIResponseHandler<decimal>> GetUserAccountBalanceByAccountNoAndUserIdRepositoryAsync(int userID,string accountNo);
         Task <APIResponseHandler<UserAccountDetail>> CreateUserAccountDetailsRepositoryAsync(UserAccountDetail userAccountDetail);
 }
